working, but keeps enter name here
