+++ lib/index.js
@@ -1,20 +1,13 @@
 require('./style.scss');
 
 var Game = require("./game");
+var MenuTyper = require("./menuTyper");
 var ScoreBoard = require("./scoreBoard");
 var canvas = document.getElementById('canvas');
 var context = canvas.getContext('2d');
 var scoreBoard = new ScoreBoard();
+var menuTyper = new MenuTyper()
 
-function populateScoreBoard(highScores) {
-  $(".score").remove();
-  $("#scoreboard-post-type").hide();
-  highScores.forEach(function(score, index){
-    var scoreInfo = score.split("~")
-    $("#scoreboard-post-type").append("<p class='score'>" + index + ". " + scoreInfo[0] + " : " + scoreInfo[1] + "</p>" )
-  })
-  typeStuff(" cat leaderboard.txt", $("#scoreboard-typer"), "scoreboard");
-}
 
 function clearStatusMenu() {
   $(".echo").remove();
@@ -25,7 +18,7 @@ function showInstructions(){
   $(".menu").hide();
   $(".instructions").show();
   $("#instructions-post-type").hide();
-  typeStuff(" cat instructions.txt", $("#instructions-typer"), "instructions");
+  menuTyper.typeStuff(" cat instructions.txt", $("#instructions-typer"), "instructions");
 }
 
 function setupPregameDisplay() {
@@ -39,7 +32,7 @@ function setupPregameDisplay() {
 
 function setupGameDisplay() {
   clearStatusMenu();
-  populateScoreBoard(scoreBoard.getHighScores());
+  // scoreBoard.populate();
   $("#restartButton").hide();
   $(".status p").remove();
   $("#status-tab").click();
@@ -48,9 +41,10 @@ function setupGameDisplay() {
 
 function setupPostgameDisplay(score) {
   $("#restartButton").show();
-  $("#scoreboard-tab").click();
   scoreBoard.getLeaderName(score);
-  populateScoreBoard(scoreBoard.getHighScores());
+    // $("#scoreboard-tab").click();
+    // // populateScoreBoard(scoreBoard.getHighScores());
+
 }
 
 function gameLoop() {
@@ -82,33 +76,6 @@ $(".tab").on('click', function(){
       showInstructions();
     }
     if ($(this).attr('target') === "scoreboard") {
-      populateScoreBoard(scoreBoard.getHighScores());
+      scoreBoard.populate();
     }
 } );
-
-function cursorAnimation() {
-    $('.new-cursor').animate({
-        opacity: 0
-    }, 'slow', 'swing').animate({
-        opacity: 1
-    }, 'slow ', 'swing');
-}
-
-
-function typeStuff(text, container, menuName) {
-  // $(".new-cursor").show();
-  setInterval (cursorAnimation, 800);
-  var textLength = 0;
-  function typer() {
-      container.html(text.substr(0, textLength++));
-      if(textLength < text.length+1) {
-          setTimeout(typer , 80);
-      } else {
-          $("#" + menuName + "-post-type").fadeIn(700);
-          $(".new-cursor").hide();
-          textLength = 0;
-          text = '';
-      }
-  }
-  typer();
-}
