+++ lib/scoreBoard.js
@@ -1,3 +1,6 @@
+var MenuTyper = require("./menuTyper");
+var menuTyper = new MenuTyper()
+
 class ScoreBoard {
   constructor() {
     this.currentHighScores = this.getHighScores();
@@ -18,11 +21,16 @@ class ScoreBoard {
       })
 
       localStorage.setItem("highScores", topScores);
+      $("#scoreboard-tab").click();
+      this.populate();
   }
 
   getLeaderName(score){
     if (score > this.lowestScore() || this.roomAvailable()) {
+      $(".score").remove();
+      this.topScoreClick();
       $(".add-leader").show();
+
       $("#leaderName").on("keydown", function(event) {
         if (event.which == 13 || event.keycode == 13) {
           var leaderName = $("#leaderName").html();
@@ -31,9 +39,17 @@ class ScoreBoard {
           $("#leaderName").html("Enter Name Here.")
         }
       }.bind(this))
-
+      return true
+    } else {
+      return false
     }
+  }
 
+  topScoreClick() {
+        $('.tab').removeClass("active-tab");
+        $('#scoreboard-tab').addClass("active-tab");
+        $('.menu').hide();
+        $('.scoreboard').show();
   }
 
   getHighScores() {
@@ -56,6 +72,16 @@ class ScoreBoard {
 
   lowestScore() {
     return this.currentHighScores[this.numHighScores() - 1].split("~")[1];
+  }
+
+  populate() {
+    $(".score").remove();
+    $("#scoreboard-post-type").hide();
+    this.currentHighScores.forEach(function(score, index){
+      var scoreInfo = score.split("~")
+      $("#scoreboard-post-type").append("<p class='score'>" + index + ". " + scoreInfo[0] + " : " + scoreInfo[1] + "</p>" )
+    })
+    menuTyper.typeStuff(" cat leaderboard.txt", $("#scoreboard-typer"), "scoreboard");
   }
 }
